# src/engine/server/docs/swagger.yaml
basePath: '{{.BasePath}}'
definitions:
...
  util.PluginConfigMap:
    properties:
      key1:
        type: string
      key2:
        type: string
    type: object
  util.Result:
    properties:
      code:
        type: integer
      data:
        items:
          type: string
        type: array
      messages:
        items:
          type: Message
        type: array
    type: object
...
host: '{{.Host}}'
info:
  contact:
    email: keith.tenzer@gmail.com
    name: Keith Tenzer
    url: http://www.keithtenzer.com
  description: |-
    APIs for managing Fossul workflows, jobs, profile, and configs
    JSON API definition can be retrieved at <a href="/api/v1/swagger/doc.json">/api/v1/swagger/doc.json</a>
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Fossul Framework Server API
  version: "1.0"
paths:
  /addConfig/{profileName}/{configName}:
    post:
      consumes:
      - application/json
      description: Add Configuration
      parameters:
      - description: config struct
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/util.Config'
          type: object
      - description: name of profile
        in: path
        name: profileName
        required: true
        type: string
      - description: name of config
        in: path
        name: configName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            string:
              description: OK
              type: string
          schema:
            $ref: '#/definitions/util.Result'
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
  /addPluginConfig/{profileName}/{configName}/{pluginName}:
    post:
      consumes:
      - application/json
      description: Add Plugin Configuration
      parameters:
      - description: config map
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/util.PluginConfigMap'
          type: object
      - description: name of profile
        in: path
        name: profileName
        required: true
        type: string
      - description: name of config
        in: path
        name: configName
        required: true
        type: string
      - description: name of plugin
        in: path
        name: pluginName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            string:
              description: OK
              type: string
          schema:
            $ref: '#/definitions/util.Result'
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
  /addProfile/{profileName}:
    get:
      consumes:
      - application/json
      description: Add Profile
      parameters:
      - description: name of profile
        in: path
        name: profileName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            string:
              description: OK
              type: string
          schema:
            $ref: '#/definitions/util.Result'
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
...
  /deleteConfig/{profileName}/{configName}:
    get:
      consumes:
      - application/json
      description: Delete Configuration
      parameters:
      - description: name of profile
        in: path
        name: profileName
        required: true
        type: string
      - description: name of config
        in: path
        name: configName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            string:
              description: OK
              type: string
          schema:
            $ref: '#/definitions/util.Result'
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
  /deleteConfigDir/{profileName}/{configName}:
    get:
      consumes:
      - application/json
      description: Delete Entire Configuration (destructive)
      parameters:
      - description: name of profile
        in: path
        name: profileName
        required: true
        type: string
      - description: name of config
        in: path
        name: configName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            string:
              description: OK
              type: string
          schema:
            $ref: '#/definitions/util.Result'
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
  /deletePluginConfig/{profileName}/{configName}/{pluginName}:
    get:
      consumes:
      - application/json
      description: Add Plugin Configuration
      parameters:
      - description: name of profile
        in: path
        name: profileName
        required: true
        type: string
      - description: name of config
        in: path
        name: configName
        required: true
        type: string
      - description: name of plugin
        in: path
        name: pluginName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            string:
              description: OK
              type: string
          schema:
            $ref: '#/definitions/util.Result'
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
  /deleteProfile/{profileName}:
    get:
      consumes:
      - application/json
      description: Delete Profile Including Configurations (destructive)
      parameters:
      - description: name of profile
        in: path
        name: profileName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            string:
              description: OK
              type: string
          schema:
            $ref: '#/definitions/util.Result'
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
...
  /getConfig/{profileName}/{configName}:
    get:
      consumes:
      - application/json
      description: Get Configuration
      parameters:
      - description: name of profile
        in: path
        name: profileName
        required: true
        type: string
      - description: name of config
        in: path
        name: configName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            string:
              description: OK
              type: string
          schema:
            $ref: '#/definitions/util.Config'
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
  /getDefaultConfig:
    get:
      consumes:
      - application/json
      description: Get Default Configuration
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            string:
              description: OK
              type: string
          schema:
            $ref: '#/definitions/util.Config'
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
  /getDefaultPluginConfig/{pluginName}:
    get:
      consumes:
      - application/json
      description: Get Default Plugin Configuration
      parameters:
      - description: name of plugin
        in: path
        name: pluginName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            string:
              description: OK
              type: string
          schema:
            type: map
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
...
  /getPluginConfig/{profileName}/{configName}/{pluginName}:
    get:
      consumes:
      - application/json
      description: Get Plugin Configuration
      parameters:
      - description: name of profile
        in: path
        name: profileName
        required: true
        type: string
      - description: name of config
        in: path
        name: configName
        required: true
        type: string
      - description: name of plugin
        in: path
        name: pluginName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            string:
              description: OK
              type: string
          schema:
            type: map
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
...
  /listConfigs/{profileName}:
    get:
      consumes:
      - application/json
      description: List Configurations
      parameters:
      - description: name of profile
        in: path
        name: profileName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            string:
              description: OK
              type: string
          schema:
            $ref: '#/definitions/util.Result'
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
  /listPluginConfigs/{profileName}/{configName}:
    get:
      consumes:
      - application/json
      description: List Plugin Configuration
      parameters:
      - description: name of profile
        in: path
        name: profileName
        required: true
        type: string
      - description: name of config
        in: path
        name: configName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            string:
              description: OK
              type: string
          schema:
            $ref: '#/definitions/util.Result'
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
  /listProfiles:
    get:
      consumes:
      - application/json
      description: List Profiles
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            string:
              description: OK
              type: string
          schema:
            $ref: '#/definitions/util.Result'
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
...
swagger: "2.0"